+++ packages/logic-core/src/index.ts
@@ -26,9 +26,11 @@ import CellCountPerZoneRule from './data/rules/cellCountPerZoneRule.js';
 import CellCountRule from './data/rules/cellCountRule.js';
 import CompletePatternRule from './data/rules/completePatternRule.js';
 import ConnectAllRule from './data/rules/connectAllRule.js';
+import CollectZonesRule from './data/rules/connectZonesRule.js';
 import ContainsShapeRule from './data/rules/containsShapeRule.js';
 import CustomRule from './data/rules/customRule.js';
 import DifferentCountPerZoneRule from './data/rules/differentCountPerZoneRule.js';
+import ExactCountPerZoneRule from './data/rules/exactCountPerZoneRule.js';
 import ForesightRule from './data/rules/foresightRule.js';
 import { allRules } from './data/rules/index.js';
 import LyingSymbolRule from './data/rules/lyingSymbolRule.js';
@@ -180,9 +182,11 @@ export {
   CellCountRule,
   CompletePatternRule,
   ConnectAllRule,
+  CollectZonesRule,
   ContainsShapeRule,
   CustomRule,
   DifferentCountPerZoneRule,
+  ExactCountPerZoneRule,
   ForesightRule,
   allRules,
   LyingSymbolRule,